moved functions from leveling to utils
